create upload-to-s3 module
